+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "@foretag/tanstack-db-surrealdb",
 	"description": "Add Offline / Local First Caching & Syncing to your SurrealDB app with TanstackDB and Loro (CRDTs)",
-	"version": "0.6.12",
+	"version": "0.7.0",
 	"files": [
 		"dist"
 	],
@@ -42,24 +42,25 @@
 		"test:integration": "bun test tests/surreal.integration.test.ts"
 	},
 	"dependencies": {
-		"loro-crdt": "^1.10.8"
+		"@tanstack/db-sqlite-persistence-core": "^0.1.9",
+		"loro-crdt": "^1.12.1"
 	},
 	"peerDependencies": {
-		"@tanstack/db": "^0.5.28",
+		"@tanstack/db": "^0.6.5",
 		"@tanstack/query-core": "^5.0.0",
 		"@tanstack/query-db-collection": "^1.0.25",
 		"@standard-schema/spec": "^1.1.0",
-		"surrealdb": "^2.0.0"
+		"surrealdb": "^2.0.3"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "^2.4.8",
+		"@biomejs/biome": "^2.4.15",
 		"@standard-schema/spec": "^1.1.0",
-		"@tanstack/db": "^0.5.33",
-		"@tanstack/query-core": "^5.95.2",
-		"@tanstack/query-db-collection": "^1.0.30",
-		"@tanstack/react-db": "^0.1.77",
-		"@tanstack/react-query": "^5.95.2",
-		"@types/bun": "^1.3.11",
+		"@tanstack/db": "^0.6.5",
+		"@tanstack/query-core": "^5.100.9",
+		"@tanstack/query-db-collection": "^1.0.36",
+		"@tanstack/react-db": "^0.1.83",
+		"@tanstack/react-query": "^5.100.9",
+		"@types/bun": "^1.3.13",
 		"surrealdb": "^2.0.3",
 		"tsup": "^8.5.1"
 	}
